feat: restructure async modules
