+++ lua/dapui/util.lua
@@ -8,24 +8,20 @@ local api = vim.api
 ---@return function
 function M.create_render_loop(render)
   local render_event = async.control.event()
-  local pending = false
 
   async.run(function()
     while true do
-      if not pending then
-        render_event.wait()
-      end
-      pending = false
+      render_event.wait()
+      render_event.clear()
       xpcall(render, function(msg)
         local traceback = debug.traceback(msg, 1)
         M.notify(("Rendering failed: %s"):format(traceback), vim.log.levels.WARN)
       end)
-      async.sleep(10)
+      async.sleep(30)
     end
   end)
 
   return function()
-    pending = true
     render_event.set()
   end
 end
