+++ scripts/gendocs.lua
@@ -665,7 +665,7 @@ minidoc.generate(
     "./lua/dapui/elements/watches.lua",
     "./lua/dapui/elements/breakpoints.lua",
     "./lua/dapui/elements/console.lua",
-    "./lua/dapui/async/base.lua",
+    "./lua/dapui/async/init.lua",
     "./lua/dapui/async/control.lua",
     "./lua/dapui/async/uv.lua",
     "./lua/dapui/async/tests.lua",
