+++ tests/unit/async/tasks_spec.lua
@@ -1,3 +1,4 @@
+local tasks = require("dapui.async.tasks")
 local async = require("dapui.async")
 local a = async.tests
 local tests = require("dapui.tests")
@@ -5,7 +6,7 @@ tests.bootstrap()
 
 describe("task", function()
   a.it("provides result in callbacks when already complete", function()
-    local task = async.run(function()
+    local task = tasks.run(function()
       return "test"
     end)
     local result
@@ -15,8 +16,9 @@ describe("task", function()
     end)
     assert.equals("test", result)
   end)
+
   a.it("provides result in callbacks", function()
-    local task = async.run(function()
+    local task = tasks.run(function()
       async.sleep(5)
       return "test"
     end)
@@ -27,8 +29,9 @@ describe("task", function()
     async.sleep(10)
     assert.equals("test", result)
   end)
+
   a.it("cancels", function()
-    local task = async.run(function()
+    local task = tasks.run(function()
       async.sleep(10)
       return "test"
     end)
@@ -38,8 +41,9 @@ describe("task", function()
     assert.same("Task was cancelled", task.error().message)
     assert.Nil(task.result())
   end)
+
   a.it("returns result if cancelled when completed", function()
-    local task = async.run(function()
+    local task = tasks.run(function()
       return "test"
     end)
     async.sleep(10)
@@ -48,23 +52,24 @@ describe("task", function()
     assert.Nil(task.error())
     assert.same("test", task.result())
   end)
+
   a.it("assigns parent task", function()
-    local current = async.current_task()
-    local task = async.run(function()
+    local current = tasks.current_task()
+    local task = tasks.run(function()
       return "test"
     end)
     assert.Not.Nil(task.parent)
     assert.equal(current, task.parent)
   end)
   it("assigns no parent task", function()
-    local task = async.run(function()
+    local task = tasks.run(function()
       return "test"
     end)
     assert.Nil(task.parent)
   end)
 
-  it("returns error in function", function()
-    local task = async.run(function()
+  a.it("returns error in function", function()
+    local task = tasks.run(function()
       error("test")
     end)
     async.sleep(10)
@@ -75,52 +80,30 @@ describe("task", function()
   end)
 
   a.it("sets error when callback errors", function()
-    local bad_wrapped = async.wrap(function()
+    local bad_wrapped = tasks.wrap(function()
       error("test")
     end, 1)
-    local task = async.run(bad_wrapped)
+    local task = tasks.run(bad_wrapped)
     assert.True(task.done())
     assert.True(vim.endswith(task.error().message, "test"))
   end)
-end)
 
-describe("async helpers", function()
-  a.it("sleep", function()
-    local start = vim.loop.now()
-    async.sleep(10)
-    local end_ = vim.loop.now()
-    assert.True(end_ - start >= 10)
+  a.it("sets name on creation", function()
+    local task1 = tasks.run(function()
+      return "test"
+    end)
+    local task2 = tasks.run(function()
+      return "test"
+    end)
+    assert.matches("Task %d+", task1.name())
+    assert.matches("Task %d+", task2.name())
   end)
 
   a.it("current task", function()
     local current
-    local task = async.run(function()
-      current = async.current_task()
+    local task = tasks.run(function()
+      current = tasks.current_task()
     end)
     assert.equal(task, current)
-  end)
-
-  a.it("wrap returns values provided to callback", function()
-    local task = async.run(async.wrap(function(_, _, cb)
-      cb(1, 2)
-    end, 3))
-
-    assert.same({ 1, 2 }, { task.result() })
-  end)
-
-  a.it("join returns results", function()
-    local worker = function()
-      async.sleep(10)
-      return 1
-    end
-
-    local workers = {}
-    for _ = 1, 10 do
-      table.insert(workers, worker)
-    end
-
-    local success, results = async.join(workers)
-    assert(success, results)
-    assert.same({ { 1 }, { 1 }, { 1 }, { 1 }, { 1 }, { 1 }, { 1 }, { 1 }, { 1 }, { 1 } }, results)
   end)
 end)
